packages/operator/uat/e2e/pyproject.toml
[project]
name = "agentspec-e2e"
version = "0.1.0"
description = "AgentSpec demo cluster end-to-end test suite"
requires-python = ">=3.11"
dependencies = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "pyyaml>=6.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["."]
# Show slowest 5 tests; verbose output by default
addopts = "-v --tb=short"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
